+++ solver/common.cpp
@@ -4,7 +4,7 @@
 namespace {
 
     void print_vector(const std::vector<std::vector<int>>& matrix) {
-        int size = matrix.size();
+        int size = (int)matrix.size();
         for (int i = 0; i < size; ++i) {
             for (int j = 0; j < size; ++j) {
                 std::cout << matrix[i][j] << " ";
